+++ src/kiezellisp-lib/compiler-helpers.cs
@@ -158,7 +158,8 @@ namespace Kiezel
 
         public static string CollectParameterInfo(params DynamicMetaObject[] args)
         {
-            return "(" + ", ".Join(args.Select(x => ToPrintString(x.Value))) + ")";
+            //return "(" + ", ".Join(args.Select(x => ToPrintString(x.Value))) + ")";
+            return "(" + ", ".Join(args.Select(x => x.Value == null ? "null" : x.Value.GetType().ToString())) + ")";
         }
 
         public static string CollectParameterInfo(DynamicMetaObject target, DynamicMetaObject[] args)